pyproject.toml
[project]
name = "finagent"
version = "0.1.0"
description = "FinAgent-Evo Research Framework"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deepagents>=0.4.12",
    "deepagents-acp>=0.0.4",
    "langchain-tavily>=0.2.17",
    "langchain-core>=1.2.22",
    "langchain-openai>=1.1.12",
    "langgraph-cli[inmem]>=0.4.19",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
    "zstandard>=0.25.0",
]
